remove Visual Studio 6 files, replace win32 by x64 compilation platform, fix compilation warnings
git-svn-id: https://svn.code.sf.net/p/carworld/code/trunk@32 f69ab8bf-2835-0410-8f28-9f2700608592

+++ src/H_Graphics.h
@@ -81,7 +81,7 @@ public:
 	static void Rotate(REAL teta, const Point<3,GLfloat> &Axe);	//rotate in rad around axe
 	static inline void Scale(REAL scale)
 		{glScalef(scale,scale,scale);}
-	static void LockArrays(int first, int count);
+	static void LockArrays(int first, size_t count);
 	static void UnlockArrays();
 	static inline void Begin(int type)
 		{glBegin((GLenum)type);}
